goto

+++ Classes/base/mc.h
@@ -35,7 +35,7 @@ namespace engine
 		virtual bool tryPlay(int times=1);
 		virtual bool isPlayEnd();
 		virtual void gotoAndStopByTime(float time, const string & _aniName="");
-		virtual void gotoAndStop(float progress, const string & _aniName="");
+		virtual void gotoAndStop(double progress, const string & _aniName="");
 
 
         virtual void stop(const string &  aniName = "");
@@ -165,7 +165,7 @@ namespace engine
 		virtual void onceMovieHandler(cocos2d::EventCustom *event);
         virtual void gotoAndStop(int cf, const string &  aniName = "");
 		virtual void gotoAndStopByTime(float time, const string & _aniName="");
-		virtual void gotoAndStop(float progress, const string & _aniName="");
+		virtual void gotoAndStop(double progress, const string & _aniName="");
 
         virtual void update(float dt = 0);
 		virtual const dragonBones::Rectangle & getRectangle();
@@ -219,7 +219,7 @@ namespace engine
 		inline virtual float getAlpha() { return  BaseNode::getAlpha(getDisplayNode()); };
 		inline virtual string getName() { return getDisplayNode()->getName(); };
 		virtual void gotoAndStopByTime(float time, const string & _aniName = "");
-		virtual void gotoAndStop(float progress, const string & _aniName = "");
+		virtual void gotoAndStop(double progress, const string & _aniName = "");
 
 		virtual float getWidth();
 		virtual float getHeight();
